Move test case loading to its own class
